Module xswag_base::diag

Types and functions for error diagnostic and printing

Structs

PrintOptions
Options for printing on the terminal. By default() everything is enabled.

Remark
Part of a Report that describes the occurrence with an optional code snippet.

Report
Describes some kind of problem or occurrence in the code. Contains one or more remarks with descriptions and separate code spans.

Enums

RemarkKind
Kinds of remarks

ReportKind
A report can either be an Error or a Warning. Still pretty similar to RemarkType -- may be merged with it in the future.

Functions

print
Pretty prints a report